// PremiereInterna-master/src/Models/IPaciente.ts
export interface IPaciente {
  id?: number;
  nome?: string;
  cpf?: string;
  passaporte?: string;
  data_nascimento?: any;
  nome_responsavel?: string;
  cpf_responsavel?: string;
  genero?: string;
  endereco?: string;
  cep?: string;
  bairro?: string;
  cidade?: string;
  estado?: any;
  telefone?: string;
  email?: string;
  plano_saude?: any;
  num_plano?: string;
  validade_plano?: any;
}